+++ CHANGELOG.md
@@ -7,6 +7,7 @@ NOTE: isort follows the [semver](https://semver.org/) versioning standard.
   - Fixed #1302: comments and --trailing-comma can generate invalid code.
   - Fixed #1293: extra new line in indented imports, when immediately followed by a comment.
   - Fixed #1304: isort 5 no longer recognises `sre_parse` as a stdlib module.
+  - Fixed #1300: add_imports moves comments following import section.
 
 ### 5.0.5 July 7, 2020
   - Fixed #1285: packaging issue with bundling tests via poetry.